refactor: refactoriza modelos de documentType y se elimina el modelo typeDocument
