--- app/Models/CtlTypeDocument.php
@@ -1,18 +1,0 @@
-<?php
-
-namespace App\Models;
-
-use Illuminate\Database\Eloquent\Model;
-use Illuminate\Database\Eloquent\Relations\HasOne;
-
-class CtlTypeDocument extends Model
-{
-    protected $table = "ctl_type_document";
-    protected $fillable = [
-        "id", "name", "description", "active"
-    ];
-
-    public function document(): HasOne {
-        return $this->hasOne(MntDocument::class);
-    }
-}
