+++ database/seeders/CtlDocumentTypeSeeder.php
@@ -6,14 +6,14 @@ use Illuminate\Database\Console\Seeds\WithoutModelEvents;
 use Illuminate\Database\Seeder;
 use Illuminate\Support\Facades\DB;
 
-class CtlTypeDocumentSeeder extends Seeder
+class CtlDocumentTypeSeeder extends Seeder
 {
     /**
      * Run the database seeds.
      */
     public function run(): void
     {
-        DB::table('ctl_type_document')->insert([
+        DB::table('ctl_document_type')->insert([
             "name"=>"DUI",
             "description"=> "Documento Único de Identidad",
             "active" => true,
