+++ database/seeders/DatabaseSeeder.php
@@ -2,6 +2,7 @@
 
 namespace Database\Seeders;
 
+use App\Models\CtlDocumentType;
 use App\Models\User;
 // use Illuminate\Database\Console\Seeds\WithoutModelEvents;
 use Illuminate\Database\Seeder;
@@ -27,7 +28,7 @@ class DatabaseSeeder extends Seeder
             CtlDistrictSeeder::class,
             CtlMaritalStatusSeeder::class,
             CtlGenderSeeder::class,
-            CtlTypeDocumentSeeder::class,
+            CtlDocumentType::class,
             MntPeopleSeeder::class,
         ]);
     }
